My page

+++ src/components/Profile.jsx
@@ -5,16 +5,19 @@ import { collection, deleteDoc, doc, getDocs, query } from 'firebase/firestore';
 import {
   DeleteBtn,
   MainColorSpan,
+  MakeWrap,
   MyPageTitle,
   ProfileTitle,
   ProfileWrap,
   VideoContainer,
   VideoTitle
 } from 'styles/StyledProfile';
+import { useNavigate } from 'react-router-dom';
 
 function Profile() {
   const [worldCupList, setWorldCupList] = useState([]);
   const userMail = JSON.parse(localStorage.getItem('email'));
+  const navigator = useNavigate();
 
   useEffect(() => {
     const fetchData = async () => {
@@ -39,8 +42,11 @@ function Profile() {
 
   const onDeleteClick = async (id) => {
     try {
-      await deleteDoc(doc(db, 'worldCupList', id));
-      setWorldCupList((prevWorldCupList) => prevWorldCupList.filter((worldCup) => worldCup.id !== id));
+      const deleteConfirm = window.confirm('삭제하시겠습니까?');
+      if (deleteConfirm) {
+        await deleteDoc(doc(db, 'worldCupList', id));
+        setWorldCupList((prevWorldCupList) => prevWorldCupList.filter((worldCup) => worldCup.id !== id));
+      }
     } catch (error) {
       console.error('Error deleting document:', error);
     }
@@ -53,30 +59,46 @@ function Profile() {
 
   return (
     <>
-      <ProfileTitle>
-        <p>
-          {emailRename(userMail)}님의 <MainColorSpan>WorldCupList</MainColorSpan>
-        </p>
-      </ProfileTitle>
-      {worldCupList?.map((worldCup) => (
-        <ProfileWrap key={worldCup.id}>
-          <div>
-            <MyPageTitle>
-              {emailRename(userMail)}님의 &nbsp;
-              <MainColorSpan>{worldCup.worldCupTitle}</MainColorSpan>
-            </MyPageTitle>
-          </div>
-          <VideoContainer>
-            <DeleteBtn onClick={() => onDeleteClick(worldCup.id)}>삭제하기</DeleteBtn>
-            {worldCup.videoList.map((video) => (
-              <div key={video.videoId}>
-                <YouTube videoId={video.videoId} />
-                <VideoTitle>{video.videoTitle}</VideoTitle>
+      {worldCupList.length <= 0 ? (
+        <>
+          <ProfileTitle>
+            <p>
+              {emailRename(userMail)}님의 <MainColorSpan>WorldCupList</MainColorSpan>
+            </p>
+          </ProfileTitle>
+          <MakeWrap>
+            <p>월드컵을 만들어 주세요!</p>
+            <button onClick={() => navigator('/makeWorldCup')}>New WorldCup Make here</button>
+          </MakeWrap>
+        </>
+      ) : (
+        <>
+          <ProfileTitle>
+            <p>
+              {emailRename(userMail)}님의 <MainColorSpan>WorldCupList</MainColorSpan>
+            </p>
+          </ProfileTitle>
+          {worldCupList?.map((worldCup) => (
+            <ProfileWrap key={worldCup.id}>
+              <div>
+                <MyPageTitle>
+                  {emailRename(userMail)}님의 &nbsp;
+                  <MainColorSpan>{worldCup.worldCupTitle}</MainColorSpan>
+                </MyPageTitle>
               </div>
-            ))}
-          </VideoContainer>
-        </ProfileWrap>
-      ))}
+              <VideoContainer>
+                <DeleteBtn onClick={() => onDeleteClick(worldCup.id)}>삭제하기</DeleteBtn>
+                {worldCup.videoList.map((video) => (
+                  <div key={video.videoId}>
+                    <YouTube videoId={video.videoId} />
+                    <VideoTitle>{video.videoTitle}</VideoTitle>
+                  </div>
+                ))}
+              </VideoContainer>
+            </ProfileWrap>
+          ))}
+        </>
+      )}
     </>
   );
 }
